So close...
This build is so close to working perfectly. The only issue is that there is that in order to ensure that the callbacks don't both access memory at the same time, I have to sacrifice our data processing rate. If I can either slow down our data speed or increase that rate by just a little, it will work.